requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
newspaper3k==0.2.8
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.0